feat: Replace standalone quest map and NFT screens with a new `arsenal` tab and introduce global type definitions.
